// notelab-client/src/connectors/gmail/src/types.ts
export type GmailMessagePartHeader = {
  name: string;
  value: string;
};

export type GmailMessagePartBody = {
  attachmentId?: string;
  data?: string;
  size?: number;
};

export type GmailMessagePart = {
  body?: GmailMessagePartBody;
  filename?: string;
  headers?: GmailMessagePartHeader[];
  mimeType?: string;
  partId?: string;
  parts?: GmailMessagePart[];
};

export type GmailMessage = {
  historyId?: string;
  id: string;
  internalDate?: string;
  labelIds?: string[];
  payload?: GmailMessagePart;
  raw?: string;
  sizeEstimate?: number;
  snippet?: string;
  threadId: string;
};